Add a Robot viewer tht uses the data from the sim.
This adds a subclass of pyglet's Window that manages the display
of the robots in a suitably sized window, a Queue wrapper that
handles the data pumping in a background thread, and modifies the
main script to use this instead of printing the data.

Also add some util functions & constants to the glutils.